+++ client/src/components/Container/Container.vue
@@ -125,6 +125,7 @@ export default {
 .list-group {
   min-width: 250px;
   margin-bottom: 1rem;
+  height: fit-content;
 
   &:last-child {
     margin-bottom: 0;