+++ src/Modules/Register/Register.js
@@ -1,6 +1,5 @@
 import '@/core'
 import '@/globals'
-import 'jquery-dynatree'
 import { vueApply, vueRegister } from '@/vue'
 import RegisterForm from './components/RegisterForm.vue'