+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ## Refactoring
 - Reduce complexity of RegionTransaction.GetJoinMessage() by session information !3242 @KoeMai
 - Remove unneeded u_loadCoords function !3277 @stcz
+- Reworked region admin page and removed XhrMethods #1412 #1414 #1416 #1417 #1181 !3193 @AntonBallmaier
 - Modernized EventControl to a Symfony controller !3260 @chriswalg
 - Refactor the nearby baskets list to vue !3294 @alex.simm
 - Rebuild quiz module #1666 !3119 @AntonBallmaier